Added validation, design, search support

+++ app/models/Client.scala
@@ -1,3 +1,3 @@
 package models
 
-case class Client(id: Int, firstName: String, age: Int, profession: String, var phoneNumbers: List[Long])
+case class Client(id: Int, firstName: String, age: Int, profession: String, var phoneNumbers: List[Int])
